app/requirements.txt
asgiref==3.5.2
attrs==21.4.0
backports.zoneinfo==0.2.1
Django==4.0.5
django-cors-headers==3.13.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
drf-spectacular-sidecar==2022.6.1
gunicorn==20.1.0
importlib-resources==5.8.0
inflection==0.5.1
jsonschema==4.6.1
numpy==1.23.0
pandas==1.4.3
psycopg2==2.9.3
PyJWT==2.4.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
zipp==3.8.0
